pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-foundry"
version = "0.1.0"
description = "Production-style boilerplate for grounded, voice-enabled, agentic AI apps"
readme = "README.md"
authors = [{ name = "AI Superjack" }]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "streamlit>=1.45.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "pymongo>=4.8.0",
  "redis>=5.0.8",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
  "langchain>=0.3.0",
  "langgraph>=0.2.0",
  "celery>=5.4.0",
  "jinja2>=3.1.4",
  "rich>=13.7.1",
  "openai>=1.66.0",
  "requests>=2.32.3",
  "python-multipart>=0.0.9",
  "pypdf",
  "python-docx"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-bdd>=7.2.0",
  "respx>=0.21.1",
  "ruff>=0.6.2",
  "mypy>=1.11.1"
]


[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "services*", "shared*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
